fix moving

+++ src/uet/oop/bomberman/Game.java
@@ -12,7 +12,7 @@ import java.awt.image.DataBufferInt;
 public class Game extends Canvas {
 
     public static final int TILES_SIZE = 16,
-            WIDTH = TILES_SIZE * (int)(31),
+            WIDTH = TILES_SIZE * 31,
             HEIGHT = 13 * TILES_SIZE;
 
     public static int SCALE = 3;
